+++ lib/dsbuntis.dart
@@ -2,3 +2,4 @@ library dsbuntis;
 
 export 'src/day.dart';
 export 'src/dsbapi.dart';
+export 'src/exceptions.dart';